Add lesson endpoint url creator
