+++ src/App/screens/Instructor/state/epics/utils/createLessonsUrl/index.js
@@ -1,0 +1,30 @@
+import devQueryString from 'query-string'
+import createQueryString from '../createQueryString'
+
+export default (lessonPage) => {
+
+  const paramNamesByEnv = process.env.REACT_APP_FAKE_API
+    ? {
+        page: '_page',
+        size: '_limit',
+      }
+    : {
+        page: 'page',
+        size: 'size',
+      }
+
+  const params = {
+    [paramNamesByEnv.page]: lessonPage.page,
+    [paramNamesByEnv.size]: lessonPage.size,
+    ...(lessonPage.states
+      ? {state: lessonPage.states}
+      : {}
+    ),
+  }
+
+  const queryString = process.env.REACT_APP_FAKE_API
+    ? `?${devQueryString.stringify(params)}`
+    : createQueryString(params)
+
+  return `${lessonPage.lessons_url}${queryString}`
+}
